requirements.txt
python-telegram-bot==20.8
psutil==5.9.8
requests==2.31.0
speedtest-cli==2.1.3